Fix some crashes.
When adding a branch, make sure that it has any characters at all before
checking if it's a variable.

Change the Router methods to be on the *Router pointer. We were having some
problems with Router methods not setting properties correctly, but this fixed
it.

Check that the Endpoint.methods map is set before trying to map methods to
Handlers.

With these changes, trout is complete enough that it can run the benchmarks.