Only consider a path a match if it's a full match.
Previously, any path that our URL was a subset of (e.g. /my/path is a
subset of /my/path/is/long) would match that URL, which is unexpected
behaviour and should be considered a bug.

This fixes things by ensuring that each path is considered a match only
if it terminates at a branch that has methods and handlers assigned for
it.

There's still potential for error here, if you have variables in your
path, and use different keys for the same endpoint with different
methods; you could match the wrong one on accident, and not find the
method.