Start having automated tests for our routing.
Let's make sure we can deliver on the promises we're making. A few small
changes made it easy to write some tests that determine what handler
gets called when a request is sent through a Router, so now we can make
sure our routing behaves the way we think it does pretty trivially.