Make priorities recursive.
It was previously possible to get a more exact match that was
lower-scoring, because scores only took into account the immediate
ancestor when scoring. This commit resolves that situation by making the
scoring recursively walk through all the ancestors and include all their
scores in the current node's score, meaning deeper nodes are always
going to be more valuable than shallower nodes.