Add support for middleware.
It's gross and messy to have lots of middleware smushed into the Handler
method calls, with lots of paren-nesting. Give the router and each
endpoint-terminator (everywhere a Handler can be registered) the ability
to register one or more middleware functions that will wrap that
Handler. Router middleware will be called for every Handler, and called
before Handler-specific middleware.