Add support for r.PathValue when built for Go 1.22.
Go 1.22 introduced pattern matching into its ServeMux, and as a result
the net/http.Request type got a PathValue method, returning the value
that matched the given key in the pattern.

We can, and should, populate the data behind that mathod when using
trout's pattern matching, so handlers can standardize on that built-in
function.